requirements.txt
aiogram==3.2.0
python-dotenv==1.0.0
asyncpg==0.29.0
loguru==0.7.2
fastapi==0.110.0
uvicorn==0.27.1
pydantic
python-multipart==0.0.9
python-jose[cryptography]==3.3.0 
Jinja2
pandas 
xlsxwriter